gradle run will now load the GUI
