+++ TBI-GUI/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Tue Oct 23 13:50:08 MST 2018
+gradle.version=4.5.1
