--- gui/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Mon Oct 22 17:31:27 MST 2018
-gradle.version=4.10.2
